+++ src/helpers/node.ts
@@ -109,6 +109,10 @@ export function isPrivate(type: Node) {
   return type.kind === SyntaxKind.PrivateKeyword;
 }
 
+export function isPublic(type: Node) {
+  return type.kind === SyntaxKind.PublicKeyword;
+}
+
 export function isStatic(type: Node) {
   return type.kind === SyntaxKind.StaticKeyword;
 }